# clients/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    name: Mr
  description: |
    Documentation OpenAPI de La Caverne Malagasy.
    Couvre l'ensemble des ressources : utilisateurs, catalogue, commandes, paiements, livraisons.
  title: La Caverne Malgache API
  version: 1.0.0
servers:
- description: Serveur de production
  url: https://api.lacaverne/v1
security:
- BearerAuth: []
...
paths:
...
  /users:
    get:
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: per_page
        required: false
        schema:
          default: 20
          type: integer
        style: form
      - explode: true
        in: query
        name: role
        required: false
        schema:
          enum:
          - admin
          - simple_user
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_users_get_200_response"
          description: Liste paginée
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Accès refusé (rôle insuffisant)
      summary: Liste des utilisateurs (admin)
      tags:
      - Users
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Utilisateur créé
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Accès refusé (rôle insuffisant)
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Données invalides
      summary: Créer un utilisateur (admin)
      tags:
      - Users
      x-content-type: application/json
      x-accepts:
      - application/json
  /users/me:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Profil retourné
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
      summary: Profil de l'utilisateur connecté
      tags:
      - Users
      x-accepts:
      - application/json
    put:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: Profil mis à jour
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Données invalides
      summary: Mise à jour du profil
      tags:
      - Users
      x-content-type: application/json
      x-accepts:
      - application/json
  /users/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Supprimé
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Accès refusé (rôle insuffisant)
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Ressource introuvable
      summary: Suppression d'un utilisateur (admin)
      tags:
      - Users
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Ressource introuvable
      summary: Détail d'un utilisateur (admin)
      tags:
      - Users
      x-accepts:
      - application/json
  /users/me/addresses:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Address"
                type: array
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
      summary: Liste des adresses de l'utilisateur connecté
      tags:
      - Addresses
      x-accepts:
      - application/json
    post:
      description: |
        Endpoint d'upsert. Si `id` est absent ou null dans le payload, une nouvelle adresse est créée. Si `id` est fourni et qu'une adresse appartenant à l'utilisateur existe, elle est mise à jour. Sinon, une nouvelle adresse est créée.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
          description: Adresse créée
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
          description: Adresse mise à jour
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
      summary: Créer ou mettre à jour une adresse
      tags:
      - Addresses
      x-content-type: application/json
      x-accepts:
      - application/json
  /users/me/addresses/{id}:
    delete:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Supprimée
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Ressource introuvable
      summary: Supprimer une adresse
      tags:
      - Addresses
      x-accepts:
      - application/json
    put:
      description: |
        Si l'adresse existe pour l'utilisateur connecté, elle est mise à jour. Sinon, une nouvelle adresse est créée et l'identifiant retourné fait foi.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
          description: Adresse mise à jour
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
          description: Adresse créée
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Ressource introuvable
      summary: Créer ou mettre à jour une adresse par identifiant
      tags:
      - Addresses
      x-content-type: application/json
      x-accepts:
      - application/json
  /users/me/addresses/{id}/default:
    put:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
          description: Adresse définie par défaut
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Token JWT manquant ou invalide
      summary: Définir comme adresse par défaut
      tags:
      - Addresses
      x-accepts:
      - application/json
...
components:
...
  schemas:
...
    User:
      example:
        firstname: firstname
        role: admin
        phone: phone
        id: 0
        email: email
        lastname: lastname
        status: status
      properties:
        id:
          type: integer
        firstname:
          type: string
        lastname:
          type: string
        email:
          format: email
          nullable: true
          type: string
        phone:
          nullable: true
          type: string
        role:
          enum:
          - admin
          - simple_user
          type: string
        status:
          type: string
      type: object
    UserUpdate:
      example:
        firstname: firstname
        phone: phone
        lastname: lastname
      properties:
        firstname:
          type: string
        lastname:
          type: string
        phone:
          type: string
      type: object
    UserCreateInput:
      description: "Fournir `email` ou `phone`, ou les deux."
      example:
        firstname: firstname
        password: password
        role: admin
        phone: phone
        email: email
        lastname: lastname
        status: status
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          format: email
          nullable: true
          type: string
        phone:
          nullable: true
          type: string
        password:
          format: password
          type: string
        role:
          enum:
          - admin
          - simple_user
          type: string
        status:
          type: string
      required:
      - firstname
      - lastname
      - password
      type: object
    Address:
      example:
        country_code: country_code
        user_id: 6
        location: location
        id: 0
        postal_code: postal_code
        is_default: false
      properties:
        id:
          type: integer
        user_id:
          type: integer
        location:
          type: string
        postal_code:
          type: string
        country_code:
          maxLength: 3
          type: string
        is_default:
          default: false
          type: boolean
      type: object
    AddressInput:
      example:
        country_code: country_code
        location: location
        id: 0
        postal_code: postal_code
        is_default: true
      properties:
        id:
          nullable: true
          type: integer
        location:
          type: string
        postal_code:
          type: string
        country_code:
          maxLength: 3
          type: string
        is_default:
          type: boolean
      required:
      - country_code
      - location
      - postal_code
      type: object
...
    PaginatedMeta:
      example:
        per_page: 1
        total: 0
        last_page: 5
        page: 6
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        last_page:
          type: integer
      type: object
    Error:
      example:
        code: 6
        message: message
      properties:
        code:
          type: integer
        message:
          type: string
      type: object
    _users_get_200_response:
      example:
        data:
        - firstname: firstname
          role: admin
          phone: phone
          id: 0
          email: email
          lastname: lastname
          status: status
        - firstname: firstname
          role: admin
          phone: phone
          id: 0
          email: email
          lastname: lastname
          status: status
        meta:
          per_page: 1
          total: 0
          last_page: 5
          page: 6
      properties:
        data:
          items:
            $ref: "#/components/schemas/User"
          type: array
        meta:
          $ref: "#/components/schemas/PaginatedMeta"
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http